Moved enums into a separate package.
