+++ AnnoMapEditor.Tests/Assets.cs
@@ -1,4 +1,5 @@
 using AnnoMapEditor.MapTemplates;
+using AnnoMapEditor.MapTemplates.Enums;
 using AnnoMapEditor.Mods.Enums;
 using AnnoMapEditor.Mods.Models;
 using AnnoMapEditor.Tests.Utils;
