+++ AnnoMapEditor/MapTemplates/Enums/IslandDifficulty.cs
@@ -2,7 +2,7 @@
 using System.Collections.Generic;
 using System.Linq;
 
-namespace AnnoMapEditor.MapTemplates
+namespace AnnoMapEditor.MapTemplates.Enums
 {
     public class IslandDifficulty
     {
