+++ AnnoMapEditor/MapTemplates/Models/RandomIslandElement.cs
@@ -1,4 +1,6 @@
 ﻿using Anno.FileDBModels.Anno1800.MapTemplate;
+using AnnoMapEditor.MapTemplates.Enums;
+using IslandType = AnnoMapEditor.MapTemplates.Enums.IslandType;
 
 namespace AnnoMapEditor.MapTemplates.Models
 {
